chore: upgrade to beta-4 dependencies (#1192)
* chore: upgrade to rust 1.71

* rebase origin/develop

* Bump dependencies across workspace

* Adjust fuel types in fuel-indexer-types

* Remove temp workarounds in indexer macro

* Bring in fuel-tx v0.35.3, thank you @xgreenx

* Fix most of explorer; optimize data decoding in indexer macro

* Clippy fixes

* Fix indexing_postgres e2e tests

* Manually convert contract IDs in indexer codegen

* Fix final type errors

* Upgrade fuel-indexer-tests

* Add missing fields to client schema

* Finish the upgrade...hopefully

* Upgrade to fuels-rs v0.45

* ra0x3 feedback

* Fix CI failures

---------

+++ packages/fuel-indexer/src/service.rs
@@ -360,7 +360,7 @@ async fn create_service_task(
 async fn get_start_block(
     conn: &mut IndexerConnection,
     manifest: &Manifest,
-) -> Result<u64, IndexerError> {
+) -> Result<u32, IndexerError> {
     match &manifest.resumable() {
         Some(resumable) => {
             let last = queries::last_block_height_for_indexer(
